feat: Add react-select dependency and update Components/card page
The code changes add the react-select dependency to the package.json file and update the Components/card page in the pages-manifest.json file. This change allows for the usage of the react-select component in the project.

These changes aim to enhance the project by adding the react-select dependency and updating the Components/card page to incorporate the new component.