chore: Update react-select dependency and remove Components/card page
The code changes update the react-select dependency in the package.json file and remove the Components/card page from the pages-manifest.json file. This change ensures that the project is using the latest version of react-select and removes the unused Components/card page.

These changes aim to improve the project by updating dependencies and removing unnecessary files.